requirements.txt
discord.py==1.7.3
ffmpeg-python==0.2.0
asyncio==3.4.3
requests==2.31.0
python-dotenv==1.0.0
pynacl==1.5.0
